+++ internal/graph/tuplemapper_test.go
@@ -1,4 +1,4 @@
-package tuplemapper
+package graph
 
 import (
 	"context"
